requirements.txt
python-telegram-bot==20.7
Pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.3
pymongo==4.6.1
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
flask==3.0.0
gunicorn==21.2.0
